requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-RESTful==0.3.9
Flask-Migrate==3.1.0
Flask-Cors==3.0.10
Flask-Bcrypt==0.7.1
flask-apispec==0.11.0
Flask-JWT==0.3.2
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.26.1
psycopg2==2.9.1
python-dotenv==0.19.0
